requirements.txt
numpy<2.0.0
paho_mqtt==2.1.0
